requirements.txt
apscheduler
discord.py
python-dotenv
psycopg2-binary
random_word